+++ codda/project/sample_base/client_build/app_build/src/main/java/kr/pe/codda/impl/message/BoardInfoListRes/BoardInfoListResDecoder.java
@@ -105,18 +105,18 @@ public final class BoardInfoListResDecoder extends AbstractMessageDecoder {
 				, null // nativeItemCharset
 				, boardInfo$2MiddleWritableObject));
 
-			boardInfo$2.setCnt((Integer)
+			boardInfo$2.setCnt((Long)
 			singleItemDecoder.getValueFromReadableMiddleObject(pathStack.peek()
 				, "cnt" // itemName
-				, kr.pe.codda.common.type.SingleItemType.INTEGER // itemType
+				, kr.pe.codda.common.type.SingleItemType.LONG // itemType
 				, -1 // itemSize
 				, null // nativeItemCharset
 				, boardInfo$2MiddleWritableObject));
 
-			boardInfo$2.setTotal((Integer)
+			boardInfo$2.setTotal((Long)
 			singleItemDecoder.getValueFromReadableMiddleObject(pathStack.peek()
 				, "total" // itemName
-				, kr.pe.codda.common.type.SingleItemType.INTEGER // itemType
+				, kr.pe.codda.common.type.SingleItemType.LONG // itemType
 				, -1 // itemSize
 				, null // nativeItemCharset
 				, boardInfo$2MiddleWritableObject));
